Add files via upload
This project connects to a Firestore database to manage data retrieval and submission for stadium entry. When a fan scans their ticket, the system retrieves the match_document_id and then fetches the ticket owners from the TicketsTo field. The corresponding images for each fan are obtained to enable face recognition. The camera initiates the face recognition process upon scanning. If the fan in front of the camera matches the retrieved photo, they are granted entry; otherwise, access is denied, ensuring only the rightful owner enters the stadium.