Add files via upload
Additionally, to prevent unauthorized entry, the system validates the date and time of the match. If a fan attempts to enter the stadium at an incorrect date or time, access is denied. This ensures that only fans with valid tickets for the scheduled match can enter the stadium, preventing misuse.